Combine create and edit mode
