+++ userInterface/Components/DefaultState.js
@@ -1,5 +1,4 @@
 const ReadMode = require("./ReadMode.js")
-const CreateMode = require("./CreateMode.js")
 const EditMode = require("./EditMode.js")
 
 class DefaultState {
@@ -12,13 +11,11 @@ class DefaultState {
 
     hideAllModes() {
         ReadMode.container.style.display = "none"
-        EditMode.container.style.display = "none"
-        CreateMode.mainCanvas.style.display = "none"
+        EditMode.mainCanvas.style.display = "none"
     }
 
     madeAllModesButtonsClickable() {
         ReadMode.button.disabled = false
-        CreateMode.button.disabled = false
         EditMode.button.disabled = false    
     }
 }
