+++ userInterface/index.html
@@ -425,24 +425,20 @@ button:active {
         </div>
 
         <div class="container">
-            <div id="createModeContainer" class="tooltip">
-                <button id="createMode">1. Create Mode</button>
-                <span class="tooltiptext">Drag your mouse to form a text rectangle</span>
-            </div>
             <div id="editModeContainer" class="tooltip">
-                <button id="editMode">2. Edit Mode</button>
+                <button id="editMode">1. Edit Mode</button>
                 <span class="tooltiptext">Edit text and delete existing textbox</span>
             </div>
             <div id="readModeContainer" class="tooltip">
-                <button id="readMode">3. Read Mode</button>
+                <button id="readMode">2. Read Mode</button>
                 <span class="tooltiptext">Hover your mouse on text box to read</span>
             </div>
             <div id="overlayContainer" class="tooltip">
-                <button id="overlayMode">4. Text Removal Mode</button>
+                <button id="overlayMode">3. Text Removal Mode</button>
                 <span class="tooltiptext">whiten out text boxes</span>
             </div>
             <div id="textOverlayContainer" class="tooltip">
-                <button id="textOverlayMode">5. Text Overlay Mode</button>
+                <button id="textOverlayMode">4. Text Overlay Mode</button>
                 <span class="tooltiptext">Overlay translation on text boxes</span>
             </div>
         </div>
